move events from MapConfiguration to Map
